requirements.txt
click==7.1.2
configparser==5.0.0
docx-mailmerge==0.5.0
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
Jinja2==2.11.2
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
openpyxl==3.0.3
sh==1.13.1
